feat(segmentation): Refactor DICOM handling and add utility functions for segmentation
