+++ src/imgtools/modules/segmentation/__init__.py
@@ -1,0 +1,8 @@
+from .segmentation import Segmentation
+from .utils import accepts_segmentations, map_over_labels
+
+__all__ = [
+    "accepts_segmentations",
+    "map_over_labels",
+    "Segmentation",
+]
